docs: Add DTS_FUNCTIONAL_SPECIFICATION.md
This document provides the functional specification for the Distributed Token Store (DTS) service, derived from the earlier technical specification.
