+++ .gitignore
@@ -1,5 +1,4 @@
 tools
-# /gen # Allow generated protobuf files to be tracked
 /ssoctl
 /sssoctl
 /ssso
