+++ Makefile
@@ -41,8 +41,8 @@ test: ## Run unit tests
 	@echo "Tests complete."
 
 # Example build target for the CLI. Adjust path to main.go if needed.
-build: proto tidy ## Build the ssoctl CLI application
+build: tidy ## Build the ssoctl CLI application
 	@echo "Building ssoctl CLI..."
-	$(GO) build -o sssoctl ./apps/sssoctl/sssoctl.go
-	$(GO) build -o ssso ./apps/ssso/ssso.go
-	@echo "ssoctl build complete. Executable: ./sssoctl and ./ssso"
+	$(GO) build -o ssoctl ./apps/sssoctl/
+	$(GO) build -o ssoctl ./apps/ssso/
+	@echo "ssoctl build complete. Executable: ./ssoctl & ./ssso"
