+++ go.mod
@@ -19,7 +19,7 @@ require (
 	go.opentelemetry.io/otel v1.36.0
 	go.uber.org/mock v0.5.2
 	golang.org/x/crypto v0.39.0
-	golang.org/x/net v0.41.0
+	golang.org/x/net v0.40.0
 	golang.org/x/oauth2 v0.25.0
 	golang.org/x/term v0.32.0
 	google.golang.org/protobuf v1.36.1
