+++ helm/ssso-backend/templates/secrets.yaml
@@ -1,0 +1,32 @@
+{{- if .Values.createSigningKeySecret }}
+apiVersion: v1
+kind: Secret
+metadata:
+  name: {{ .Values.signingKeySecretName }}
+  labels:
+    {{- include "ssso-backend.labels" . | nindent 4 }}
+type: Opaque
+data:
+  private.pem: {{ .Values.signingKeyPrivatePem | b64enc }}
+{{- end }}
+
+{{/*
+You might want to add a secret for MongoDB credentials if deploying MongoDB
+and not using the Bitnami subchart's way of handling secrets.
+Example for a custom MongoDB deployment (not using Bitnami subchart which handles its own secrets):
+
+{{- if and .Values.mongodb.enabled .Values.mongodb.auth.enabled (not .Values.mongodb.auth.existingSecret) }}
+apiVersion: v1
+kind: Secret
+metadata:
+  name: {{ include "ssso-backend.fullname" . }}-mongodb-auth
+  labels:
+    {{- include "ssso-backend.labels" . | nindent 4 }}
+type: Opaque
+data:
+  mongodb-root-password: {{ .Values.mongodb.auth.rootPassword | b64enc }}
+  mongodb-username: {{ .Values.mongodb.auth.username | b64enc }}
+  mongodb-password: {{ .Values.mongodb.auth.password | b64enc }}
+  mongodb-database: {{ .Values.mongodb.auth.database | b64enc }}
+{{- end }}
+*/}}
